feat: decompose model router into focused modules
